:tada: Introduced the  websocket class
